UI : Can identify or create an account.
I've created a new activity that can be used to connect or to create an account.
The email is checked, the password must be more that 6 characters and if this is a sign in, the two password must be the same.
All Firebase function have been prepared.
If the user clicked on the text before the button, the 2nd password is shown up and vice-versa.
  
TODO : 
- Show password with a button
- Change activity when login is sucessful
- Check firebase error (email already in use ...)